Extract viewer and metadata pane into a live component
